Manually revert 5154641
Turns out that a generic interface for rendering libraries
isn't that good of an idea. Each library has its quirks and it's hard
to handle all of them and easy to break something.
The previous version had a broken rectangular clipping
and it was very hard to fix specifically because of a generic interface.